+++ android/app/src/main/java/com/mulkkam/data/remote/model/response/notifications/NotificationsResponse.kt
@@ -1,4 +1,4 @@
-package com.mulkkam.data.remote.model.response.notification
+package com.mulkkam.data.remote.model.response.notifications
 
 import com.mulkkam.domain.model.notification.Notification
 import kotlinx.serialization.SerialName
