+++ android/app/src/main/java/com/mulkkam/data/remote/service/VersionsService.kt
@@ -1,6 +1,6 @@
 package com.mulkkam.data.remote.service
 
-import com.mulkkam.data.remote.model.response.MinimumVersionResponse
+import com.mulkkam.data.remote.model.response.versions.MinimumVersionResponse
 import retrofit2.http.GET
 
 interface VersionsService {
