+++ android/app/src/main/java/com/mulkkam/data/repository/NotificationRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.mulkkam.data.repository
 import com.mulkkam.data.remote.model.error.toDomain
 import com.mulkkam.data.remote.model.error.toResponseError
 import com.mulkkam.data.remote.model.request.notification.ActiveCaloriesBurnedRequest
-import com.mulkkam.data.remote.model.response.notification.toDomain
+import com.mulkkam.data.remote.model.response.notifications.toDomain
 import com.mulkkam.data.remote.service.NotificationsService
 import com.mulkkam.domain.model.notification.Notification
 import com.mulkkam.domain.model.result.MulKkamResult
